✨ Now we have organized the code and implemented a search filter hook
