--- src/Playground.js
@@ -1,13 +1,0 @@
-
-import { Outlet } from 'react-router-dom';
-
-const Playground = () => {
-
-  return(
-    <div className="plays">
-      <Outlet />
-    </div>
-  )
-}
-
-export default Playground;
