+++ src/common/index.js
@@ -1,6 +1,6 @@
 import PageNotFound from './404/PageNotFound';
-import Footer from './Footer';
-import Header from './Header';
+import Footer from './footer/Footer';
+import Header from './header/Header';
 import Home from './home/Home';
 
 export {
